+++ src/main/java/de/dm/intellij/liferay/schema/LiferayJsonSchemaProviderFactory.java
@@ -25,7 +25,8 @@ public class LiferayJsonSchemaProviderFactory implements JsonSchemaProviderFacto
             new LiferayFragmentCollectionSchemaFileProvider(),
             new LiferayFragmentConfigurationSchemaFileProvider(),
             new LiferayFragmentFragmentSchemaFileProvider(),
-            new ClientExtensionYamlSchemaFileProvider()
+            new ClientExtensionYamlSchemaFileProvider(),
+            new LiferayRestConfigSchemaProvider()
         );
     }
 }